pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "PollutionTracker"
version = "1.0.0"
description = "A pollution tracking system for monitoring live, historical, and weather data using Python and fastAPI."
authors = ["Suman Sunuwar <jaycessunuwar@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"
faker = "^33.0.0"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"

[tool.poetry.group.dev.dependencies]
alembic = "^1.14.0"

